rename: ops_gpu to ops_cl
